Added enhanced test cases
