+++ rbac-service/commons/src/main/java/org/rbac/service/commons/resources/ApplicationResource.java
@@ -3,22 +3,22 @@ package org.rbac.service.commons.resources;
 import org.rbac.service.commons.annotations.Authorize;
 import org.rbac.service.commons.models.Permission;
 
-@Authorize(resource = "TEST", permission= Permission.ALL)
+@Authorize(resource = "TEST", permission = Permission.ALL)
 public class ApplicationResource {
-    @Authorize(resource = "TEST", permission= Permission.ALL)
-    public String getResult(){
+    @Authorize(resource = "TEST", permission = Permission.ALL)
+    public String getResult() {
         return null;
     }
 
 
     //              EXTRA CODE FOR DEMONSTRATION (can be removed after usage end)
-    @Authorize(resource = "TEST Database", permission= Permission.READ)
-    public String getUser(){
+    @Authorize(resource = "TEST Database", permission = Permission.READ)
+    public String getUser() {
         return "read-only";
     }
 
-    @Authorize(resource = "TEST Production", permission= Permission.WRITE)
-    public String getUser1(){
+    @Authorize(resource = "TEST Production", permission = Permission.WRITE)
+    public String getUser1() {
         return "write-only";
     }
